[HOPSWORKS-1208] Reorder installation of wit-widget
[HOPSWORKS-1208] Fix LD_LIBRARY_PATH when installing Tensorflow ROCm

[HOPSWORKS-1208] Skip installing tfx tfma if Tf-ROCm

+++ recipes/default.rb
@@ -272,6 +272,7 @@ for python in python_versions
     #yes | ${CONDA_DIR}/envs/${ENV}/bin/pip install pyspark==#{node['pyspark']['version']}
     yes | ${CONDA_DIR}/envs/${ENV}/bin/pip install hops-apache-beam==#{node['conda']['beam']['python']['version']}
 
+    yes | ${CONDA_DIR}/envs/${ENV}/bin/pip install --no-cache-dir --upgrade witwidget
     yes | ${CONDA_DIR}/envs/${ENV}/bin/pip uninstall tensorflow
     yes | ${CONDA_DIR}/envs/${ENV}/bin/pip uninstall tensorboard
     yes | ${CONDA_DIR}/envs/${ENV}/bin/pip uninstall tensorflow-estimator
@@ -287,7 +288,7 @@ for python in python_versions
     # If system is setup for rocm already or we are installing it
     else
       export ROCM=#{node['rocm']['install']}
-      if [ -f /opt/rocm/bin/rocminfo ] || [$ROCM == "true"]  ; then
+      if [ -f /opt/rocm/bin/rocminfo ] || [ $ROCM == "true" ]  ; then
         TENSORFLOW_LIBRARY_SUFFIX="-rocm"
       fi
     fi
@@ -409,8 +410,7 @@ for python in python_versions
 
       yes | ${CONDA_DIR}/envs/${ENV}/bin/pip install --no-cache-dir --upgrade hdfscontents urllib3 requests pandas
 
-      # Install wit-widget
-      yes | ${CONDA_DIR}/envs/${ENV}/bin/pip install --no-cache-dir --upgrade witwidget
+      # Install wit-widget JupyterLab extension
       source /usr/local/nvm/nvm.sh
       nvm use 10.16.0
       ${CONDA_DIR}/envs/${ENV}/bin/jupyter labextension install --no-build wit-widget
@@ -469,10 +469,15 @@ for python in python_versions
 
     code <<-EOF
       set -e
-      export PATH=$PATH:$HADOOP_HOME/bin
-      #Instal TensorFlow Extended Model Analysis extension
-        ${CONDA_DIR}/envs/${ENV}/bin/jupyter nbextension install --py --sys-prefix --symlink tensorflow_model_analysis
-        ${CONDA_DIR}/envs/${ENV}/bin/jupyter nbextension enable --py --sys-prefix tensorflow_model_analysis
+      # Tensorflow-ROCm is currently problematic with Tfx TFMA
+      if [ ! -f /opt/rocm/bin/rocminfo ] && [ #{node['rocm']['install']} != "true" ];
+      then
+            export PATH=$PATH:$HADOOP_HOME/bin
+
+            #Install TensorFlow Extended Model Analysis extension
+            ${CONDA_DIR}/envs/${ENV}/bin/jupyter nbextension install --py --sys-prefix --symlink tensorflow_model_analysis
+            ${CONDA_DIR}/envs/${ENV}/bin/jupyter nbextension enable --py --sys-prefix tensorflow_model_analysis
+      fi
     EOF
   end
